% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countbart.R
\name{count_bart}
\alias{count_bart}
\title{Fit log-linear BART for count regression models.}
\usage{
count_bart(
  y,
  x,
  x_zero = x,
  x_pos = x,
  offset = NULL,
  nburn,
  nsim,
  nthin = 1,
  update_interval = 100,
  ntree = 250,
  a0 = NA,
  nztree = 100,
  z_conc = 3.5/sqrt(2),
  sd_val = 2 * sd(y),
  base = 0.95,
  power = 2,
  nu = 3,
  lambda = NULL,
  sigq = 0.9,
  sighat = NULL,
  kappa_a = 5,
  kappa_b = 3,
  kappa_prop_sd = 0.21,
  count_model = "poisson",
  debug = FALSE,
  randeff_design = matrix(1),
  randeff_variance_component_design = matrix(1),
  randeff_scales = 1,
  randeff_df = 3
)
}
\arguments{
\item{y}{Response variable (must be vector of non-negative integers).}

\item{x}{Design matrix for the log-linear predictor, f(x).}

\item{x_zero}{Design matrix for zero-tree, which gives the odds a zero is due to the zero-inflated point mass component. Defaults to x.}

\item{x_pos}{Design matrix for positive-tree, which gives the odds a zero is NOT due to the zero-inflated point mass component. Defaults to x.}

\item{offset}{Offset value; can be a scalar or vector. If 'NULL', an offset of 1 will be used.}

\item{nburn}{Number of burn-in MCMC iterations.}

\item{nsim}{Number of MCMC iterations to save after burn-in.}

\item{nthin}{Save every nthin'th MCMC iterate. The total number of MCMC iterations will be nsim*nthin + nburn.}

\item{update_interval}{Print status every update_interval MCMC iterations.}

\item{ntree}{Number of trees in log-linear predictor component.}

\item{a0}{Concentration parameter. 'NA' gives default estimate from Murray (2022).}

\item{nztree}{Number of trees in zero-inflated point mass probability component.}

\item{z_conc}{Concentration parameter for zero-inflated log-odds prior.}

\item{sd_val}{SD(mu(x)) marginally at any covariate value (or its prior median if use_muscale=TRUE).}

\item{base}{Base for tree prior on trees (see details).}

\item{power}{Power for the tree prior on trees.}

\item{nu}{Degrees of freedom in the chisq prior on \eqn{sigma^2}.}

\item{lambda}{Scale parameter in the chisq prior on \eqn{sigma^2}.}

\item{sigq}{Calibration quantile for the chisq prior on \eqn{sigma^2}.}

\item{sighat}{Calibration estimate for the chisq prior on \eqn{sigma^2}.}

\item{kappa_a}{Shape parameter (alpha) for kappa prior (beta prime distribution); default is 5.}

\item{kappa_b}{Shape parameter (beta) for kappa prior (beta prime distribution); default is 3.}

\item{kappa_prop_sd}{Standard deviation of kappa's Metropolis-Hasting proposal distribution.}

\item{count_model}{Specifies the count model; takes values "poisson", "nb", "zipoisson", or "zinb".}

\item{debug}{XX}

\item{randeff_design}{XX}

\item{randeff_variance_component_design}{XX}

\item{randeff_scales}{XX}

\item{randeff_df}{XX}
}
\value{
A list with posterior estimates of \eqn{\log(E(Y_i | x_i))}, posterior samples of trees,
and posterior estimates of dispersion parameter \eqn{\kappa} (for negative binomial models).
}
\description{
Fit log-linear BART for count regression models.
}
\details{
Fits a generalized version of the log-linear BART model (Murray 2021): For a response
variable y and covariates x,
\deqn{y_i ~ Pr(Y_i = y_i | \mu(x_i))}
where \eqn{Pr(Y_i = y_i | \mu(x_i))} is the PMF from a Poisson or negative binomial regression 
model (or their zero-inflated variants). Furthermore, \eqn{\mu(x_i) = \mu_{0i} f(x_i)}, for fixed
offset \eqn{\mu_{0i}} and BART prior over log-linear predictor function \eqn{log(f(x_i))}. 

Some notes:
\itemize{
   \item x (and x_zero and x_pos) must be numeric matrices. See e.g. the makeModelMatrix function in the
   dbarts package for appropriately constructing a design matrix from a data.frame
   \item The prior for leaf parameter \eqn{\lambda} is a mixture of generalized inverse Gaussian (GIG)
   distributions, with hyperparameters \eqn{c} and \eqn{d}. By default, \eqn{c} and \eqn{d} are chosen
   so that \eqn{var(\lambda) = a^2_{0} / m}, where \eqn{a_0} is a concentration parameter and \eqn{m}
   denotes the number of trees in \eqn{f} (see Proposition 4.1 from Murray (2021) for details). For 
   the log-linear predictor function, \eqn{a_0} is chosen such that \eqn{Pr(f(x_i) \leq y^{*}) \approx 0.975}
   marginally, where \eqn{y^{*}} is the 0.95 quantile of the empirical distribution of y. For zero-inflated 
   odds components \eqn{f^{(0)}(x_i)} and \eqn{f^{(1)}(x_i)}, \eqn{a_0 = 3.5 / \sqrt{2}}.
   \item By default, \eqn{\kappa} (the negative binomial dispersion parameter) has a beta prime prior with 
   shape parameters \eqn{\alpha = 5} and \eqn{\beta = 3}.
}
}
\references{
Jared S. Murray (2021). Log-Linear Bayesian Additive Regression Trees for Multinomial Logistic and Count Regression Models.
 https://doi.org/10.1080/01621459.2020.1813587.
}
