% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_dummies}
\alias{get_dummies}
\title{Construct (redundant) dummies from a factor}
\usage{
get_dummies(grouping_var, original_colnames = TRUE)
}
\arguments{
\item{grouping_var}{A factor defining the groups}

\item{original_colnames}{If true, use the original factor levels as column names. Otherwise "Group" will be pre-pended to the column names of the dummy matrix}
}
\value{
A matrix of dummy variables
}
\description{
Construct (redundant) dummies from a factor
}
