% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_fit.R
\name{get_forest_fit}
\alias{get_forest_fit}
\title{Extract fits/coefficients from multibart forests}
\usage{
get_forest_fit(mb_forest, X, Z = NULL)
}
\arguments{
\item{mb_forest}{A multibart forest object}

\item{X}{Covariate matrix}

\item{Z}{Basis functions corresponding to X}
}
\value{
If Z is null, a (basis dimension) x n x (number of MCMC iterations) array
of coefficients. If Z is not null an n x (number of MCMC iterations) matrix
of fitted values
}
\description{
Extract fits/coefficients from multibart forests
}
