% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_random_intercepts_slopes_posteriors}
\alias{get_random_intercepts_slopes_posteriors}
\title{Get posteriors for random effects/parameters in a varying intercepts/slopes model}
\usage{
get_random_intercepts_slopes_posteriors(bcf_fit, randeff_setup)
}
\arguments{
\item{bcf_fit}{}

\item{randeff_setup}{}
}
\value{

}
\description{
Get posteriors for random effects/parameters in a varying intercepts/slopes model
}
