% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forest.R
\name{make_forest}
\alias{make_forest}
\title{make_forest}
\usage{
make_forest(tree_samples, tree_strings, shift = 0, scale = 1)
}
\arguments{
\item{tree_samples}{}

\item{tree_strings}{}

\item{shift}{}

\item{scale}{}
}
\value{

}
\description{
make_forest
}
