% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mb_forest.R
\name{print.mb_forest}
\alias{print.mb_forest}
\title{Print a multibart forest}
\usage{
\method{print}{mb_forest}(x)
}
\arguments{
\item{x}{A multibart forest}
}
\value{

}
\description{
Print a multibart forest
}
