% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{random_intercepts}
\alias{random_intercepts}
\title{Construct metadata for BCF with random intercepts}
\usage{
random_intercepts(groups, original_colnames = TRUE)
}
\arguments{
\item{groups}{A factor variable defining the groups}

\item{original_colnames}{If true, use the original factor levels as column names for the matrix of dummies for the groups. Otherwise "Group" will be pre-pended to the column names of the dummy matrix}
}
\value{

}
\description{
Construct metadata for BCF with random intercepts
}
